Pruebas Unitarias creadas para todos los metodos actuales

+++ src/main/java/com/example/netpolix/model/Video.java
@@ -60,5 +60,5 @@ public class Video {
     private Integer idTemporada; // Change to Integer to allow null values
 
     @Column(name = "calificacion")
-    private int calificacion;
+    private double calificacion;
 }
